refactor: move helper code to lib directory
