+++ .eslintignore
@@ -41,9 +41,7 @@ app/main.prod.js
 app/main.prod.js.map
 app/renderer.prod.js
 app/renderer.prod.js.map
-app/style.css
-app/style.css.map
-app/utils/bech32.js
+app/lib/utils/bech32.js
 dist
 dll
 main.js
