+++ app/components/Activity/TransactionModal.js
@@ -8,7 +8,7 @@ import FaAngleDown from 'react-icons/lib/fa/angle-down'
 import Isvg from 'react-inlinesvg'
 import paperPlane from 'icons/paper_plane.svg'
 import link from 'icons/link.svg'
-import { blockExplorer } from 'utils'
+import { blockExplorer } from 'lib/utils'
 
 import Value from 'components/Value'
