+++ app/lib/lnd/config.js
