+++ app/lib/lnd/index.js
@@ -1,7 +1,7 @@
 import config from './config'
-import lightning from './lib/lightning'
-import walletUnlocker from './lib/walletUnlocker'
-import isLndRunning from './lib/util'
+import lightning from './lightning'
+import walletUnlocker from './walletUnlocker'
+import isLndRunning from './util'
 import subscribe from './subscribe'
 import methods from './methods'
 import walletUnlockerMethods from './walletUnlockerMethods'
