+++ app/lib/lnd/methods/channelController.js
