+++ app/lib/lnd/methods/index.js
