+++ app/lib/lnd/methods/invoicesController.js
