+++ app/lib/lnd/methods/networkController.js
