+++ app/lib/lnd/methods/paymentsController.js
