+++ app/lib/lnd/methods/peersController.js
