+++ app/lib/lnd/methods/walletController.js
