+++ app/lib/lnd/walletUnlockerMethods/index.js
