+++ app/lib/utils/bech32.js
