+++ app/routes/activity/components/components/Transaction/Transaction.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Moment from 'react-moment'
-import { btc } from 'utils'
+import { btc } from 'lib/utils'
 
 import Value from 'components/Value'
 import styles from '../Activity.scss'
